cleaned support for amd64 and non-amd64 platforms
